refactored route folders, split home body into separate file, added presave button to home page
