+++ package.json
@@ -9,6 +9,8 @@
     "start": "next start"
   },
   "dependencies": {
+    "@emotion/react": "^11.4.0",
+    "@emotion/styled": "^11.3.0",
     "framer-motion": "^3.3.0",
     "next": "10.0.6",
     "react": "17.0.1",
