+++ pagecomponents/presave.js
@@ -1,0 +1,3 @@
+export default function PreSave() {
+  return <button>Click here</button>
+}
